+++ tests/BaseTestCase.php
@@ -2,6 +2,8 @@
 
 namespace Boscho87\tests;
 
+use Boscho87\ChangelogChecker\FileManager\File;
+use Boscho87\ChangelogChecker\FileManager\FileInterface;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -9,4 +11,11 @@ use PHPUnit\Framework\TestCase;
  */
 class BaseTestCase extends TestCase
 {
+    protected const CHANGELOG_FILE_PATH = __DIR__ . '/MockFiles/changelog-examples/KeepaChangelog.md';
+
+
+    protected function getTestMockFile(): FileInterface
+    {
+        return new File(self::CHANGELOG_FILE_PATH);
+    }
 }