+++ tests/Checkers/AbstractCheckerTest.php
@@ -2,8 +2,6 @@
 
 namespace Boscho87\tests\Checkers;
 
-use Boscho87\ChangelogChecker\FileManager\File;
-use Boscho87\ChangelogChecker\FileManager\FileInterface;
 use Boscho87\tests\BaseTestCase;
 
 /**
@@ -11,11 +9,4 @@ use Boscho87\tests\BaseTestCase;
  */
 abstract class AbstractCheckerTest extends BaseTestCase
 {
-    protected const CHANGELOG_FILE_PATH = __DIR__ . '/../MockFiles/changelog-examples/KeepaChangelog.md';
-
-
-    protected function getTestMockFile(): FileInterface
-    {
-        return new File(self::CHANGELOG_FILE_PATH);
-    }
 }